+++ mycodo/utils/influx.py
@@ -14,9 +14,11 @@ from mycodo.config import INFLUXDB_PASSWORD
 from mycodo.config import INFLUXDB_PORT
 from mycodo.config import INFLUXDB_USER
 from mycodo.databases.models import Output
+from mycodo.databases.models import Unit
 from mycodo.inputs.sensorutils import convert_units
 from mycodo.mycodo_client import DaemonControl
 from mycodo.utils.database import db_retrieve_table_daemon
+from mycodo.utils.system_pi import add_custom_units
 
 logger = logging.getLogger("mycodo.influxdb")
 
@@ -279,6 +281,8 @@ def read_influxdb_function(
     :param end_str: End time, in influxdb format
     :type end_str: str
     """
+    test_unit_channel(unit, channel)
+
     client = InfluxDBClient(
         INFLUXDB_HOST, INFLUXDB_PORT, INFLUXDB_USER, INFLUXDB_PASSWORD,
         INFLUXDB_DATABASE, timeout=5)
@@ -351,6 +355,8 @@ def read_influxdb_list(unique_id, unit, channel,
     :param end_str: End time, in influxdb format
     :type end_str: str
     """
+    test_unit_channel(unit, channel)
+
     raw_data = None
 
     client = InfluxDBClient(
@@ -419,6 +425,8 @@ def read_influxdb_single(unique_id, unit, channel,
     :param value: What kind of measurement to return (e.g. LAST, SUM, MIN, MAX, etc.)
     :type value: str
     """
+    test_unit_channel(unit, channel)
+
     client = InfluxDBClient(
         INFLUXDB_HOST, INFLUXDB_PORT, INFLUXDB_USER, INFLUXDB_PASSWORD,
         INFLUXDB_DATABASE, timeout=5)
@@ -626,6 +634,8 @@ def write_influxdb_value(
     :param timestamp: If supplied, this timestamp will be used in the influxdb
     :type timestamp: datetime object
     """
+    test_unit_channel(unit, None)
+
     client = InfluxDBClient(
         INFLUXDB_HOST, INFLUXDB_PORT, INFLUXDB_USER, INFLUXDB_PASSWORD,
         INFLUXDB_DATABASE, timeout=5)
@@ -700,3 +710,12 @@ def write_influxdb_list(data, unique_id):
                 "submitted for writing: {data}. Exception: {err}".format(
                     data=data, err=e))
             return 1
+
+
+def test_unit_channel(unit, channel):
+    """ Test if unit and channel exists """
+    if unit not in add_custom_units(Unit.query.all()):
+        raise Exception("Unit ID not found")
+    if channel is not None:
+        if not isinstance(channel, int) or channel < 0:
+            raise Exception("Channel must be an integer greater or equal to 0")