+++ mycodo/mycodo_flask/api/measurement.py
@@ -5,9 +5,9 @@ import traceback
 
 import flask_login
 from flask_accept import accept
-from flask_restplus import Resource
-from flask_restplus import abort
-from flask_restplus import fields
+from flask_restx import Resource
+from flask_restx import abort
+from flask_restx import fields
 
 from mycodo.mycodo_flask.api import api
 from mycodo.mycodo_flask.api import default_responses