+++ mycodo/mycodo_flask/api/sql_schema_fields.py
@@ -1,5 +1,5 @@
 # coding=utf-8
-from flask_restplus import fields
+from flask_restx import fields
 
 from mycodo.mycodo_flask.api import api